bridgePerform
