+++ Sources/Engine/State/SKKEvent.h
@@ -24,36 +24,9 @@
 #define SKKEvent_h
 
 #include <sstream>
-#import <AquaSKKEngine/GenericStateMachine.h>
 #import <swift/bridging>
-
-// キー入力イベント
-enum {
-    SKK_NULL = statemachinecxx_sourceforge_jp::USER_EVENT, // 無効なイベント
-    SKK_JMODE,                                             // Ctrl-J
-    SKK_ENTER,                                             // Ctrl-M
-    SKK_CANCEL,                                            // Ctrl-G
-    SKK_BACKSPACE,                                         // Ctrl-H
-    SKK_DELETE,                                            // Ctrl-D
-    SKK_TAB,                                               // Ctrl-I
-    SKK_PASTE,                                             // Ctrl-Y
-    SKK_LEFT,                                              // ←
-    SKK_RIGHT,                                             // →
-    SKK_UP,                                                // ↑
-    SKK_DOWN,                                              // ↓
-    SKK_CHAR,                                              // その他全てのキー入力
-    SKK_PING,                                              // CTRL-L(内部状態問い合わせ)
-    SKK_UNDO,                                              // CTRL-/
-    SKK_ASCII_MODE,                                        // ASCII モード
-    SKK_HIRAKANA_MODE,                                     // ひらかなモード
-    SKK_KATAKANA_MODE,                                     // カタカナモード
-    SKK_JISX0201KANA_MODE,                                 // 半角カナモード
-    SKK_JISX0208LATIN_MODE,                                // 全角英数モード
-    SKK_YES,                                               // 仮想イベント
-    SKK_NO,                                                // 仮想イベント
-    SKK_ON,                                                // 仮想イベント
-    SKK_OFF                                                // 仮想イベント
-};
+#import <AquaSKKEngine/GenericStateMachine.h>
+#import <AquaSKKEngine/SKKEventID.h>
 
 // SKK_CHAR 属性
 enum {
@@ -83,8 +56,6 @@ enum {
     PseudoHandled, // 処理は行うが「未処理」とする
     CapsLock       // CapsLock
 };
-
-#import <AquaSKKEngine/SKKBridgedState.h>
 
 // イベントパラメータ
 class SKKEvent {
@@ -232,10 +203,6 @@ public:
         buf << "code=0x" << std::hex << (unsigned)code << ", " << attr();
 
         return buf.str();
-    }
-
-    SKKBridgedEventID getBridgedEventID() const SWIFT_COMPUTED_PROPERTY {
-        return static_cast<SKKBridgedEventID>(id);
     }
 };
