+++ Sources/Engine/utility/AquaSKKEngine-Preamble.h
@@ -20,12 +20,12 @@ namespace AquaSKKEngine {
 #import <AquaSKKEngine/SKKCandidateWindowBridge.h>
 #import <AquaSKKEngine/SKKCompleter.h>
 #import <AquaSKKEngine/SKKEvent.h>
-#import <AquaSKKEngine/SKKState.h>
 #import <AquaSKKEngine/SKKInputContext.h>
 #import <AquaSKKEngine/SKKInputEnvironment.h>
 #import <AquaSKKEngine/SKKInputQueueObserver.h>
 #import <AquaSKKEngine/SKKOkuriListener.h>
 #import <AquaSKKEngine/SKKSelectorBuddy.h>
+#import <AquaSKKEngine/SKKState.h>
 #import <AquaSKKEngine/SKKStateMachine.h>
 
 #endif /* AquaSKKEngine_Preamble_h */
